Update on line 176 balances[currentWinner]
I think we can map the balance of msg.sender to msg.value. So the mapping is more accurate.
eg. balances[currentWinner] = msg.value;
This way the mapping of balance of msg.sender will always be the amount they send in ethers
I apologise in advance if I was wrong